ml-service/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# ML Libraries
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
joblib==1.3.2

# Image Processing
opencv-python==4.9.0.80
Pillow==10.2.0
imagehash==4.3.1

# Text Processing
nltk==3.8.1

# EXIF Metadata
piexif==1.1.3

# HTTP Client
requests==2.31.0
aiohttp==3.9.1

# Environment Variables
python-dotenv==1.0.0

# Geospatial
geopy==2.4.1
